redfishpower: refactor with command state
Problem: Redfishpower uses a flag to indicate which "state" an on/off
power command is currently in.  In the future, we may add an additional
"state" to the protocol.  It will become inconvenient to continue to
use flags.

Internally, convert to use a "state" variable with a series of
state transitions.