redfishpower: support config of connection timeout
Problem: In large clusters, it is common for hardware to be missing for
maintenance or repair.  This makes it likely that "pm -q" will always
timeout after 5 seconds, making the use of the command annoying.  While
support of --message-timeout makes this more tolerable, we don't want
the general message timeout to be made too low.

Support a new --connection-timeout option, which will configure the libcurl
timeout specific to connection activities.  Default this to 2 seconds, a
value that appears to balance "reasonable length" and "not too annoying".